+++ src/Resolution.ts
@@ -145,6 +145,19 @@ export default class Resolution {
   }
 
   /**
+   * Resolve a chat id from the domain record
+   * @param domain - domain name to be resolved
+   * @throws ResolutionError with code RecordNotFound
+   * @returns A promise that resolves in chatId
+   */
+  async chatId(domain: string): Promise<string> {
+    domain = this.prepareDomain(domain);
+    const method = this.getNamingMethodOrThrow(domain);
+    return await method.chatId(domain);
+  }
+
+
+  /**
    * Resolves the IPFS hash configured for domain records on ZNS
    * @param domain - domain name
    * @throws ResolutionError